+++ etherman/types.go
@@ -1,7 +1,6 @@
 package etherman
 
 import (
-	"math/big"
 	"time"
 
 	"github.com/0xPolygonHermez/zkevm-node/etherman/smartcontracts/proofofefficiency"
@@ -24,7 +23,7 @@ type Block struct {
 // GlobalExitRoot struct
 type GlobalExitRoot struct {
 	BlockNumber       uint64
-	GlobalExitRootNum *big.Int
+	Timestamp         time.Time
 	MainnetExitRoot   common.Hash
 	RollupExitRoot    common.Hash
 	GlobalExitRoot    common.Hash
@@ -54,6 +53,7 @@ type VerifiedBatch struct {
 	BlockNumber uint64
 	BatchNumber uint64
 	Aggregator  common.Address
+	StateRoot   common.Hash
 	TxHash      common.Hash
 }
