+++ state/globalexitroot.go
@@ -1,7 +1,7 @@
 package state
 
 import (
-	"math/big"
+	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 )
@@ -9,7 +9,7 @@ import (
 // GlobalExitRoot struct
 type GlobalExitRoot struct {
 	BlockNumber       uint64
-	GlobalExitRootNum *big.Int
+	Timestamp         time.Time
 	MainnetExitRoot   common.Hash
 	RollupExitRoot    common.Hash
 	GlobalExitRoot    common.Hash
