+++ state/pgstatestorage_test.go
@@ -2,6 +2,7 @@ package state_test
 
 import (
 	"context"
+	"math"
 	"math/big"
 	"testing"
 	"time"
@@ -160,6 +161,91 @@ func TestAddAndGetSequences(t *testing.T) {
 	require.Equal(t, uint64(7), sequences[1].FromBatchNumber)
 	require.Equal(t, uint64(7), sequences[0].ToBatchNumber)
 	require.Equal(t, uint64(8), sequences[1].ToBatchNumber)
+
+	require.NoError(t, dbTx.Commit(ctx))
+}
+
+func TestAddGlobalExitRoot(t *testing.T) {
+	// Init database instance
+	initOrResetDB()
+
+	ctx := context.Background()
+	tx, err := testState.BeginStateTransaction(ctx)
+	require.NoError(t, err)
+	block := &state.Block{
+		BlockNumber: 1,
+		BlockHash:   common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		ParentHash:  common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		ReceivedAt:  time.Now(),
+	}
+	err = testState.AddBlock(ctx, block, tx)
+	assert.NoError(t, err)
+	globalExitRoot := state.GlobalExitRoot{
+		BlockNumber:       1,
+		Timestamp:         time.Now(),
+		MainnetExitRoot:   common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		RollupExitRoot:    common.HexToHash("0x30a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+		GlobalExitRoot:    common.HexToHash("0x40a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+	}
+	err = testState.AddGlobalExitRoot(ctx, &globalExitRoot, tx)
+	require.NoError(t, err)
+	exit, _, err := testState.GetLatestGlobalExitRoot(ctx, math.MaxInt64, tx)
+	require.NoError(t, err)
+	err = tx.Commit(ctx)
+	require.NoError(t, err)
+	assert.Equal(t, globalExitRoot.BlockNumber, exit.BlockNumber)
+	assert.Equal(t, globalExitRoot.Timestamp.Unix(), exit.Timestamp.Unix())
+	assert.Equal(t, globalExitRoot.MainnetExitRoot, exit.MainnetExitRoot)
+	assert.Equal(t, globalExitRoot.RollupExitRoot, exit.RollupExitRoot)
+	assert.Equal(t, globalExitRoot.GlobalExitRoot, exit.GlobalExitRoot)
+}
+
+func TestVerifiedBatch(t *testing.T) {
+	initOrResetDB()
+
+	ctx := context.Background()
+	dbTx, err := testState.BeginStateTransaction(ctx)
+	require.NoError(t, err)
+
+	block := &state.Block{
+		BlockNumber: 1,
+		BlockHash:   common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		ParentHash:  common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		ReceivedAt:  time.Now(),
+	}
+	err = testState.AddBlock(ctx, block, dbTx)
+	assert.NoError(t, err)
+	//require.NoError(t, tx.Commit(ctx))
+
+	lastBlock, err := testState.GetLastBlock(ctx, dbTx)
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(1), lastBlock.BlockNumber)
+
+	_, err = testState.PostgresStorage.Exec(ctx, "INSERT INTO state.batch (batch_num) VALUES (1)")
+
+	require.NoError(t, err)
+	virtualBatch := state.VirtualBatch{
+		BlockNumber: 1,
+		BatchNumber: 1,
+		TxHash:      common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+	}
+	err = testState.AddVirtualBatch(ctx, &virtualBatch, dbTx)
+	require.NoError(t, err)
+	expectedVerifiedBatch := state.VerifiedBatch{
+		BlockNumber: 1,
+		BatchNumber: 1,
+		StateRoot:   common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f2"),
+		Aggregator:  common.HexToAddress("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		TxHash:      common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+	}
+	err = testState.AddVerifiedBatch(ctx, &expectedVerifiedBatch, dbTx)
+	require.NoError(t, err)
+
+	// Step to create done, retrieve it
+
+	actualVerifiedBatch, err := testState.GetVerifiedBatch(ctx, 1, dbTx)
+	require.NoError(t, err)
+	require.Equal(t, expectedVerifiedBatch, *actualVerifiedBatch)
 
 	require.NoError(t, dbTx.Commit(ctx))
 }
